requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
pydantic>=2.0
python-dotenv
psycopg[binary]
bcrypt==3.2.2
passlib==1.7.4
redis>=5.0
celery>=5.3
python-jose[cryptography]
email-validator
